requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
Brotli==1.1.0
CacheControl==0.14.2
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
clip==1.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cssselect2==0.7.0
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
distro==1.9.0
docstring_parser==0.16
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.10.0
filelock==3.13.1
filetype==1.2.0
firebase-admin==6.2.0
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.56.0
fsspec==2024.6.1
ftfy==6.3.1
google-api-core==2.24.0
google-api-python-client==2.160.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.75.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0rc1
googlemaps==4.10.0
grpc-google-iam-v1==0.14.0
grpcio==1.68.1
grpcio-status==1.67.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.1.0
nest_asyncio==1.6.0
networkx==3.3
numpy==2.2.2
openai==1.61.1
opencv-python==4.11.0.86
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.5
pickleshare==0.7.5
pillow==10.4.0
pip==25.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
proto-plus==1.26.0
protobuf==5.29.2
psutil==6.1.1
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydyf==0.11.0
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
pypdf==5.2.0
pypdfium2==4.30.0
pyphen==0.17.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2025.1
PyVirtualDisplay==3.0
pywin32==307
PyYAML==6.0.2
pyzmq==26.2.1
RapidFuzz==3.12.1
regex==2024.11.6
requests==2.32.3
requirement.txt==0.1.0
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
selenium==4.28.1
sendgrid==6.11.0
sentence-transformers==3.4.1
setuptools==70.2.0
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack_data==0.6.3
starkbank-ecdsa==2.2.0
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.0
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchsummary==1.5.1
torchvision==0.21.0+cu118
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.2
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vertexai==1.71.1
wcwidth==0.2.13
weasyprint==64.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.21.0
zopfli==0.2.3.post1
